requirements.txt
# --- Core UI Framework ---
gradio==4.31.4

# --- AI / LLM Integration ---
groq==0.4.2
python-dotenv==1.0.0
httpx==0.27.0
requests==2.31.0

# --- Audio / Speech ---
gTTS==2.5.1
pydub==0.25.1
soundfile==0.12.1

# --- Data Processing ---
numpy==1.24.3
pandas==2.0.3
matplotlib==3.8.3

# --- Auth / Security ---
python-jose==3.3.0
passlib==1.7.4
python-multipart>=0.0.9




# --- Async / Server ---
aiohttp==3.9.3

# --- Validation / Typing ---
pydantic==2.5.2
typing-extensions==4.8.0
